Refactor VideoTube workspace into modular views
